refactor: classify local upload staging boundary
