+++ apps/agor-daemon/src/utils/upload-staging-defaults.ts
@@ -1,0 +1,2 @@
+export const DEFAULT_UPLOAD_TTL_MS = 24 * 60 * 60 * 1000;
+export const DEFAULT_UPLOAD_MAX_BYTES = 50 * 1024 * 1024;
